Version 2.1.0 Released
* New Annotator: MongoStemming uses a gazetteer and stemming to perform
a pseudo-fuzzy match and find gazetter terms in different tenses and
plurals
* New Cleaner: MergeAdjacent will merge adjacent entities of the same
type
* New Content Extractor: CsvContentExtractor splits CSV fields into
content and metadata
* New Collection Reader: LineReader will read a single file into
multiple documents by line
* New REST API to get configuration parameters for components (e.g.
annotators)
* Significant changes to the way gazetteer annotators work, including
changing from RadixTrees to MultiMaps and implementing the Aho-Corasick
algorithm, resulting in performance improvements for large gazetteers in
the order of 100s
* Lots of bug fixes and minor improvements

+++ baleen/baleen-annotators/src/main/java/uk/gov/dstl/baleen/annotators/regex/internals/MoneyRegex.java
@@ -46,7 +46,7 @@ public class MoneyRegex extends AbstractRegexAnnotator<Money> {
 				if(matcher.group(20).toLowerCase().startsWith("p")){
 					money.setCurrency("GBP");
 				}
-				money.setAmount(Float.parseFloat(matcher.group(19))/100);
+				money.setAmount(Double.parseDouble(matcher.group(19))/100);
 			}else if(matcher.group(25) != null){	//e.g. 42 euros
 				if(matcher.group(30).toLowerCase().startsWith("euro")){
 					money.setCurrency("EUR");
